# api/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления ПВЗ и приемкой товаров",
        "title": "PVZ service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pvz": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of PVZ points with optional filters. Requires JWT-Token with Employee or Moderator role.",
                "tags": [
                    "PVZ"
                ],
                "summary": "List PVZ points",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of PVZ points",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pvz.PVZListResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new PVZ. Requires JWT-Token with Employee role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVZ"
                ],
                "summary": "Create PVZ",
                "operationId": "CreatePVZ",
                "parameters": [
                    {
                        "description": "PVZ creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvz.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PVZ successfully created",
                        "schema": {
                            "$ref": "#/definitions/pvz.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "City not found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "409": {
                        "description": "Pvz with this id already exists",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/pvz/{pvzID}/close_last_reception": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Close the last reception for a given PVZ ID. Requires JWT-Token with Employee role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVZ"
                ],
                "summary": "Close Last Reception",
                "operationId": "CloseLastReception",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PVZ ID",
                        "name": "pvzID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully closed last reception",
                        "schema": {
                            "$ref": "#/definitions/reception.CloseLastReceptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing PVZ ID",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "No open reception found for this PVZ",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/receptions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new reception entry. Requires JWT-Token with Employee role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reception"
                ],
                "summary": "Create Reception",
                "operationId": "CreateReception",
                "parameters": [
                    {
                        "description": "Reception creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reception.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Reception successfully created",
                        "schema": {
                            "$ref": "#/definitions/reception.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "PVZ not found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pvz.CreateRequest": {
            "type": "object",
            "required": [
                "city",
                "id",
                "registrationDate"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                }
            }
        },
        "pvz.CreateResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                }
            }
        },
        "pvz.PVZListResponse": {
            "type": "object",
            "properties": {
                "pvz": {
                    "$ref": "#/definitions/pvz.PvzResponse"
                },
                "receptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvz.ReceptionsWithProduct"
                    }
                }
            }
        },
        "pvz.ProductsResponse": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "receptionId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pvz.PvzResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                }
            }
        },
        "pvz.ReceptionsResponse": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pvzId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "pvz.ReceptionsWithProduct": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pvz.ProductsResponse"
                    }
                },
                "reception": {
                    "$ref": "#/definitions/pvz.ReceptionsResponse"
                }
            }
        },
        "reception.CloseLastReceptionResponse": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pvzID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "reception.CreateRequest": {
            "type": "object",
            "required": [
                "pvzID"
            ],
            "properties": {
                "pvzID": {
                    "type": "string"
                }
            }
        },
        "reception.CreateResponse": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pvzID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Bearer authentication. The API enforces role-based authorization with two roles: moderator and employee.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}